backend/requirements.txt
# FastAPI dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
websockets>=11.0.2
pydantic>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0

# Flask dependencies
flask>=2.3.0
requests>=2.28.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
eventlet>=0.33.0

# AI Agent dependencies
langgraph>=0.0.10
transformers>=4.30.0
torch>=2.0.0
bitsandbytes>=0.41.0  # For 8-bit quantization
accelerate>=0.21.0    # For device mapping
einops>=0.6.1         # Common requirement for LLMs
safetensors>=0.3.1    # For model loading

# Testing dependencies
unittest2>=1.1.0
pytest>=7.3.1